requirements.txt
numpy==1.18.5
tensorflow-gpu==2.3.1
scikit-learn==0.24.0
keras==2.4.3
matplotlib==3.1.3
scipy==1.5.2
pandas==1.1.3
orange3==3.27.1
baycomp==1.0.2